pyproject.toml
[tool.poetry]
name = "aicenter"
version = "0.1.0"
description = ""
authors = ["kitUIN <kulujun@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
django-comment-migrate = "^0.1.7"
django-cors-headers = "^4.6.0"
django-filter = "^24.3"
django-ranged-response = "^0.2.0"
django-redis = "^5.4.0"
django-restql = "^0.15.4"
django-timezone-field = "^7.0"
djangorestframework = "^3.15.2"
channels = "^4.2.0"
channels-redis = "^4.2.1"
uvicorn = "^0.32.0"
gunicorn = "^23.0.0"
gevent = "^24.10.3"
celery = "^5.5.0rc2"
django-celery-beat = "^2.7.0"
django-celery-results = "^2.5.1"
drf-nested-routers = "^0.94.1"
django-concurrency = "^2.6.0"
mysqlclient = "^2.2.5"
drf-yasg = "^1.21.8"
label-studio-sdk = "^1.0.7"
fastapi = "^0.115.6"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
